Remove doc about deprecated values for the SSL option 'verify'
In the old SSL implentation, 0,1 and 2 were the possible values for the 'verify'
option. For the new one, only verify_peer and verify_none are officialy
supported and documentated. Old values are still supported for backward
compatiliby, but not documented.

We do the same in Yaws. Old values are translated into new ones. So if their
support is removed from the SSL application in a futur Erlang/OTP release, Yaws
will still work. Documentation has been updated to remove info about the old
values.