Move Erlang compatibily checks from the configure to a module
Now, the new module yaws_dynopts does all tests on the Erlang/OTP release that
were done before in the configure. So it is possible to compile Yaws with an
Erlang/OTP release and run it with another, all features should be dynamically
detected.

Because these checks can be expensive, when Yaws is started, we generate and
compile a static version of yaws_dynopts.